Use double instead of float for generated C code
Double precision floating point numbers are well supported by most
architectures, and C will generate reasonable code for them.

This creates a disrepancy between the single precision floats used in
the RISC virtual machine and the C code, but genuine double support for
RISC may not matter much of the time, and a RISC register can't hold a
64 bit value.

+++ examples/cairo.c
@@ -24,12 +24,12 @@ void Cairo_WriteToPNG(Cairo_Context *ctx, const char *fileName, int len) {
   cairo_surface_write_to_png(c->surface, fileName);
 }
 
-void Cairo_MoveTo(Cairo_Context *ctx, float x, float y) {
+void Cairo_MoveTo(Cairo_Context *ctx, double x, double y) {
   Cairo_Context_Native *c = (Cairo_Context_Native*)ctx;
   cairo_move_to(c->cr, x, y);
 }
 
-void Cairo_LineTo(Cairo_Context *ctx, float x, float y) {
+void Cairo_LineTo(Cairo_Context *ctx, double x, double y) {
   Cairo_Context_Native *c = (Cairo_Context_Native*)ctx;
   cairo_line_to(c->cr, x, y);
 }
